+++ crates/benchmark-suite/benches/jsonschema.rs
@@ -3,7 +3,7 @@ use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use serde_json::Value;
 
 fn bench_build(c: &mut Criterion, name: &str, schema: &Value) {
-    c.bench_function(&format!("jsonschema/{}/build", name), |b| {
+    c.bench_function(&format!("jsonschema/{name}/build"), |b| {
         b.iter_with_large_drop(|| jsonschema::validator_for(schema).expect("Valid schema"))
     });
 }