+++ crates/jsonschema/benches/keywords.rs
@@ -3,7 +3,7 @@ use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use serde_json::Value;
 
 fn bench_keyword_build(c: &mut Criterion, name: &str, schema: &Value) {
-    c.bench_function(&format!("keyword/{}/build", name), |b| {
+    c.bench_function(&format!("keyword/{name}/build"), |b| {
         b.iter_with_large_drop(|| jsonschema::validator_for(schema).expect("Valid schema"))
     });
 }
@@ -11,7 +11,7 @@ fn bench_keyword_build(c: &mut Criterion, name: &str, schema: &Value) {
 fn bench_keyword_is_valid(c: &mut Criterion, name: &str, schema: &Value, instance: &Value) {
     let validator = jsonschema::validator_for(schema).expect("Valid schema");
     c.bench_with_input(
-        BenchmarkId::new(format!("keyword/{}", name), "is_valid"),
+        BenchmarkId::new(format!("keyword/{name}"), "is_valid"),
         instance,
         |b, instance| {
             b.iter(|| {
@@ -24,7 +24,7 @@ fn bench_keyword_is_valid(c: &mut Criterion, name: &str, schema: &Value, instanc
 fn bench_keyword_validate(c: &mut Criterion, name: &str, schema: &Value, instance: &Value) {
     let validator = jsonschema::validator_for(schema).expect("Valid schema");
     c.bench_with_input(
-        BenchmarkId::new(format!("keyword/{}", name), "validate"),
+        BenchmarkId::new(format!("keyword/{name}"), "validate"),
         instance,
         |b, instance| {
             b.iter(|| {