+++ crates/jsonschema/src/error.rs
@@ -98,7 +98,9 @@ pub enum ValidationErrorKind {
     /// Unexpected properties.
     AdditionalProperties { unexpected: Vec<String> },
     /// The input value is not valid under any of the schemas listed in the 'anyOf' keyword.
-    AnyOf,
+    AnyOf {
+        context: Vec<Vec<ValidationError<'static>>>,
+    },
     /// Results from a [`fancy_regex::RuntimeError::BacktrackLimitExceeded`] variant when matching
     BacktrackLimitExceeded { error: fancy_regex::Error },
     /// The input value doesn't match expected constant.
@@ -146,7 +148,9 @@ pub enum ValidationErrorKind {
     /// The given schema is valid under more than one of the schemas listed in the 'oneOf' keyword.
     OneOfMultipleValid,
     /// The given schema is not valid under any of the schemas listed in the 'oneOf' keyword.
-    OneOfNotValid,
+    OneOfNotValid {
+        context: Vec<Vec<ValidationError<'static>>>,
+    },
     /// When the input doesn't match to a pattern.
     Pattern { pattern: String },
     /// Object property names are invalid.
@@ -228,15 +232,21 @@ impl<'a> ValidationError<'a> {
             schema_path: location,
         }
     }
-    pub(crate) const fn any_of(
+    pub(crate) fn any_of(
         location: Location,
         instance_path: Location,
         instance: &'a Value,
+        context: Vec<Vec<ValidationError<'a>>>,
     ) -> ValidationError<'a> {
         ValidationError {
             instance_path,
             instance: Cow::Borrowed(instance),
-            kind: ValidationErrorKind::AnyOf,
+            kind: ValidationErrorKind::AnyOf {
+                context: context
+                    .into_iter()
+                    .map(|errors| errors.into_iter().map(|error| error.to_owned()).collect())
+                    .collect(),
+            },
             schema_path: location,
         }
     }
@@ -602,15 +612,21 @@ impl<'a> ValidationError<'a> {
             schema_path: location,
         }
     }
-    pub(crate) const fn one_of_not_valid(
+    pub(crate) fn one_of_not_valid(
         location: Location,
         instance_path: Location,
         instance: &'a Value,
+        context: Vec<Vec<ValidationError<'a>>>,
     ) -> ValidationError<'a> {
         ValidationError {
             instance_path,
             instance: Cow::Borrowed(instance),
-            kind: ValidationErrorKind::OneOfNotValid,
+            kind: ValidationErrorKind::OneOfNotValid {
+                context: context
+                    .into_iter()
+                    .map(|errors| errors.into_iter().map(|error| error.to_owned()).collect())
+                    .collect(),
+            },
             schema_path: location,
         }
     }
@@ -765,13 +781,13 @@ fn write_quoted_list(f: &mut Formatter<'_>, items: &[impl fmt::Display]) -> fmt:
     let mut iter = items.iter();
     if let Some(item) = iter.next() {
         f.write_char('\'')?;
-        write!(f, "{}", item)?;
+        write!(f, "{item}")?;
         f.write_char('\'')?;
     }
     for item in iter {
         f.write_str(", ")?;
         f.write_char('\'')?;
-        write!(f, "{}", item)?;
+        write!(f, "{item}")?;
         f.write_char('\'')?;
     }
     Ok(())
@@ -801,11 +817,11 @@ impl fmt::Display for ValidationError<'_> {
                 let mut iter = array.iter().skip(*limit);
 
                 if let Some(item) = iter.next() {
-                    write!(f, "{}", item)?;
+                    write!(f, "{item}")?;
                 }
                 for item in iter {
                     f.write_str(", ")?;
-                    write!(f, "{}", item)?;
+                    write!(f, "{item}")?;
                 }
 
                 write_unexpected_suffix(f, array.len() - limit)
@@ -815,12 +831,12 @@ impl fmt::Display for ValidationError<'_> {
                 write_quoted_list(f, unexpected)?;
                 write_unexpected_suffix(f, unexpected.len())
             }
-            ValidationErrorKind::AnyOf => write!(
+            ValidationErrorKind::AnyOf { context: _ } => write!(
                 f,
                 "{} is not valid under any of the schemas listed in the 'anyOf' keyword",
                 self.instance
             ),
-            ValidationErrorKind::OneOfNotValid => write!(
+            ValidationErrorKind::OneOfNotValid { context: _ } => write!(
                 f,
                 "{} is not valid under any of the schemas listed in the 'oneOf' keyword",
                 self.instance
@@ -831,7 +847,7 @@ impl fmt::Display for ValidationError<'_> {
                 self.instance
             ),
             ValidationErrorKind::Constant { expected_value } => {
-                write!(f, "{} was expected", expected_value)
+                write!(f, "{expected_value} was expected")
             }
             ValidationErrorKind::ContentEncoding { content_encoding } => {
                 write!(
@@ -927,7 +943,7 @@ impl fmt::Display for ValidationError<'_> {
             }
             ValidationErrorKind::PropertyNames { error } => error.fmt(f),
             ValidationErrorKind::Required { property } => {
-                write!(f, "{} is a required property", property)
+                write!(f, "{property} is a required property")
             }
             ValidationErrorKind::MultipleOf { multiple_of } => {
                 write!(f, "{} is not a multiple of {}", self.instance, multiple_of)
@@ -955,13 +971,13 @@ impl fmt::Display for ValidationError<'_> {
                 let mut iter = types.iter();
                 if let Some(t) = iter.next() {
                     f.write_char('"')?;
-                    write!(f, "{}", t)?;
+                    write!(f, "{t}")?;
                     f.write_char('"')?;
                 }
                 for t in iter {
                     f.write_str(", ")?;
                     f.write_char('"')?;
-                    write!(f, "{}", t)?;
+                    write!(f, "{t}")?;
                     f.write_char('"')?;
                 }
                 Ok(())
@@ -994,12 +1010,12 @@ impl fmt::Display for MaskedValidationError<'_, '_, '_> {
                 write_quoted_list(f, unexpected)?;
                 write_unexpected_suffix(f, unexpected.len())
             }
-            ValidationErrorKind::AnyOf => write!(
+            ValidationErrorKind::AnyOf { context: _ } => write!(
                 f,
                 "{} is not valid under any of the schemas listed in the 'anyOf' keyword",
                 self.placeholder
             ),
-            ValidationErrorKind::OneOfNotValid => write!(
+            ValidationErrorKind::OneOfNotValid { context: _ } => write!(
                 f,
                 "{} is not valid under any of the schemas listed in the 'oneOf' keyword",
                 self.placeholder
@@ -1010,7 +1026,7 @@ impl fmt::Display for MaskedValidationError<'_, '_, '_> {
                 self.placeholder
             ),
             ValidationErrorKind::Constant { expected_value } => {
-                write!(f, "{} was expected", expected_value)
+                write!(f, "{expected_value} was expected")
             }
             ValidationErrorKind::ContentEncoding { content_encoding } => {
                 write!(
@@ -1110,7 +1126,7 @@ impl fmt::Display for MaskedValidationError<'_, '_, '_> {
             }
             ValidationErrorKind::PropertyNames { error } => error.fmt(f),
             ValidationErrorKind::Required { property } => {
-                write!(f, "{} is a required property", property)
+                write!(f, "{property} is a required property")
             }
             ValidationErrorKind::MultipleOf { multiple_of } => {
                 write!(
@@ -1144,13 +1160,13 @@ impl fmt::Display for MaskedValidationError<'_, '_, '_> {
                 let mut iter = types.iter();
                 if let Some(t) = iter.next() {
                     f.write_char('"')?;
-                    write!(f, "{}", t)?;
+                    write!(f, "{t}")?;
                     f.write_char('"')?;
                 }
                 for t in iter {
                     f.write_str(", ")?;
                     f.write_char('"')?;
-                    write!(f, "{}", t)?;
+                    write!(f, "{t}")?;
                     f.write_char('"')?;
                 }
                 Ok(())